requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kubernetes==29.0.0
pyyaml==6.0.1
pytest==7.4.3
httpx==0.24.1
jinja2>=3.1.2
